Allow WithConfigFileFlag and WithConfigFile to be meaningfully used together
Prior to this commit, WithConfigFileFlag and WithConfigFile could not be
used together, as WithConfigFile would always take precedence.

It's conceivable that a tool author would want a tool user to be able to
specify a config file if wanted, but fallback to a default if one was
not provided. This change in semantics allows for this behavior. It's
unlikely people used and wanted WithConfigFileFlag and WithConfigFile
together prior to this change, so it's reasonably safe to add.