Provide more contextualized usage messages
Usage messages should include the names and flags for all commands
from the root to the terminal command. Previously usage functions
could only access the terminal command's data, which made it difficult
to automatically generate a fully contextuatlized usage message. The
programmer was forced to manually copy the information into each
subcommands usage function.

This commit changes Command.Parse to collect the sequence of commands
found from the beginning of the parse to the terminal command found.
It also changes the signature of UsageFunc to accept the sequence of
commands found during parsing. These changes provide implementations
of UsageFunc with the data necessary to automatically generate more
detailed usage messages with less work required by programmers.

Note: The UsageFunc signature change is not backwards compatible and
requires increasing the major version of the module.