Type alias for runtime-constrained flag set `any`
Having an `any` in the signature for `Parse` is unsatisfying because it
is actually a runtime-constrained value, and `any` does not convey this
information.

In lieu of support for "union type" general interfaces which allow the
intermixing of basic interfaces with concrete types, a type alias to
`any` with a descriptive name at least places the constraint
front-and-center.